% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_from_coords.R
\name{blocks_from_coords}
\alias{blocks_from_coords}
\title{Find Census Blocks from Lat/Lng Coordinates}
\usage{
blocks_from_coords(filename, year)
}
\arguments{
\item{filename}{the path to a file which contains latitude ("lat"), longitude ("lng"), and state ("state"), with those names}

\item{year}{the census year that shapefiles will be drawn from for tracts (2000, 2010, 2015); defaults to 2016}
}
\description{
This function allows you to geocode a dataset of coordinates into census blocks and isolates a GISJOIN variable that enables simple
and easy merging with Census and American Community Survey data
}
\details{
Warning: Census block shapefiles are massive and require a dependable and strong internet connection. The block files of some larger states can exceed 0.5GBs
}
\examples{
tracts_from_coords(filename = "your_filename.csv", year = 2010)

}
\keyword{blocks}
\keyword{census}
