% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfs.R
\name{compute_tfs}
\alias{compute_tfs}
\title{Compute the Tract Frequency Spectrum (TFS)}
\usage{
compute_tfs(bin_mat, n_sample = 100)
}
\arguments{
\item{bin_mat}{Binary matrix. Rows correspond to bin, columns to sampled chromosomes/individuals.}

\item{n_sample}{Integer. Number of sampled chromosomes/individuals. Ideally it should be equivalent to the number of columns, but it is not if some individuals do not carry any tracts.}
}
\value{
Vector. It corresponds to the TFS.
}
\description{
Given a binary matrix the function computes the Tract Frequency Spectrum (TFS).
}
