% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_adj_mat}
\alias{get_adj_mat}
\title{Get adjacency matrix}
\usage{
get_adj_mat(tracts_gr, chrom, pos)
}
\arguments{
\item{chrom}{}
}
\value{
Matrix.
}
\description{
Given a GRanges object containing tracts and a genomic position, the function returns the corresponding adjacency matrix for the tracts
intersecting that genomic location. The matrix is symmetric and the row and column names correspond to the ID of the individuals carrying the tracts.
Two individuals sharing one recombination breakpoints will have an entry equal to 1, two individuals sharing two rec breakpoints (i.e., they share the same tract) have an entry equal to 2.
}
