% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_bin_mat_sites}
\alias{get_bin_mat_sites}
\title{Get binary matrix of tracts for recombination sites approach - EMPIRICAL and SIMULATION}
\usage{
get_bin_mat_sites(tracts_gr)
}
\arguments{
\item{tracts_gr}{}
}
\value{
Binary
}
\description{
Get binary matrix of tracts for recombination sites approach - EMPIRICAL and SIMULATION
}
