% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_bin_mat_subtracts}
\alias{get_bin_mat_subtracts}
\title{Get binary matrix of tracts for subtracts approach - EMPIRICAL and SIMULATION}
\usage{
get_bin_mat_subtracts(tracts_gr)
}
\arguments{
\item{tracts_gr}{}
}
\description{
Get binary matrix of tracts for subtracts approach - EMPIRICAL and SIMULATION
}
