% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_bin_mat_subtracts_sim}
\alias{get_bin_mat_subtracts_sim}
\title{Get binary matrix of tracts for subtracts approach - SIMULATION}
\usage{
get_bin_mat_subtracts_sim(tracts_gr)
}
\arguments{
\item{tracts_gr}{}
}
\description{
Get binary matrix of tracts for subtracts approach - SIMULATION
}
