% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_bin_mat_unique_emp}
\alias{get_bin_mat_unique_emp}
\title{Get binary matrix of tracts for uniqueness approach - EMPIRICAL}
\usage{
get_bin_mat_unique_emp(tracts_gr)
}
\arguments{
\item{tracts_gr}{}
}
\description{
Get binary matrix of tracts for uniqueness approach - EMPIRICAL
}
