% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_bin_mat_windows_sim}
\alias{get_bin_mat_windows_sim}
\title{Get binary matrix of tracts for windows approach - SIMULATION}
\usage{
get_bin_mat_windows_sim(
  tracts_gr,
  window_size = 50000,
  step_size = 50000,
  len_chr = NULL,
  ncores = 1
)
}
\arguments{
\item{ncores}{}
}
\description{
Get binary matrix of tracts for windows approach - SIMULATION
}
