% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractR.r
\name{getTracts}
\alias{getTracts}
\title{Wrapper for the U.S. Census Batch Geocoder}
\usage{
getTracts(data, unique_id, address, city, state, postalcode)
}
\arguments{
\item{data}{The data to be geocoded}

\item{unique_id}{The unique identifier required for each row in the data}

\item{address}{The street address to be geocoded}

\item{city}{The city for the address}

\item{state}{The state for the address}

\item{postalcode}{The postal code for the address}
}
\description{
This function allows you to submit batch geocoders
}
\keyword{api}
\keyword{census,}
\keyword{geocding,}
