% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts_from_coords.R
\name{tracts_from_coords}
\alias{tracts_from_coords}
\title{Find Census Tracts from Lat/Lng Coordinates}
\usage{
tracts_from_coords(filename, year)
}
\arguments{
\item{filename}{the path to a file which contains latitude ("lat"), longitude ("lng"), and state ("state"), with those names}

\item{year}{the census year that shapefiles will be drawn from for tracts (2000, 2010, 2015); defaults to 2016}
}
\description{
This function allows you to geocode a dataset of coordinates into census tracts and isolates a GISJOIN variable that enables simple
and easy merging with Census and American Community Survey data
}
\examples{
tracts_from_coords(filename = "your_filename.csv", year = 2010)

}
\keyword{census}
\keyword{tracts}
