% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_bin_mat_windows_emp}
\alias{get_bin_mat_windows_emp}
\title{Get binary matrix of tracts for windows approach - EMPIRICAL}
\usage{
get_bin_mat_windows_emp(tracts_gr, window_size = 50000, step_size = 50000)
}
\arguments{
\item{step_size}{}
}
\description{
Get binary matrix of tracts for windows approach - EMPIRICAL
}
